Finalized Health task #50
Moved the PlayerIdle asset back into LevelDesign folder and out of the scene since we added character model tracking to the camera. Also got rid of the backup box collider under the platform so the player wouldn't fall since we have no need for this anymore and actually want the player to potentially fall to its death.